Updated cleanup script
