+++ backend/app/crud/secret.py
@@ -161,3 +161,17 @@ def verify_secret_password(secret: models.Secret, password: str) -> bool:
     if secret.password_hash is None:
         return True
     return verify_password(password, secret.password_hash)
+
+
+def delete_expired_secrets(db: Session) -> int:
+    """
+    Delete all secrets that have passed their expiry time or have been revoked.
+
+    Returns:
+        int: number of deleted rows.
+    """
+    from datetime import datetime, timezone
+    now = datetime.now(timezone.utc)
+    result = (
+        db.query(models.Secret).filter(models.Secret.expires_at <= now).delete(synchronize_session=False))
+    return result
