+++ docker-compose.yml
@@ -25,8 +25,8 @@ services:
       sh -c "
         uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload &
         while true; do
-          echo '[cleanup] Running refresh token cleanup...';
-          python -m app.maintenance.cleanup_refresh_tokens;
+          echo '[cleanup] Running maintenance cleanup...';
+          python -m app.maintenance.cleanup_db;
           echo '[cleanup] Sleeping for 24h...';
           sleep 86400;
         done
